refactoring utilities, and checking for duplicate in both immutables and mutables.
