+++ utilities/property/duplicate.go
@@ -1,12 +1,12 @@
-package properties
+package property
 
 import (
 	"github.com/persistenceOne/persistenceSDK/schema/types"
 )
 
-func Duplicate(properties types.Properties) bool {
+func Duplicate(propertyList []types.Property) bool {
 	propertyIDMap := map[types.ID]bool{}
-	for _, property := range properties.GetList() {
+	for _, property := range propertyList {
 		if _, ok := propertyIDMap[property.GetID()]; ok {
 			return true
 		}
