+++ utilities/property/duplicate_test.go
@@ -1,4 +1,4 @@
-package properties
+package property
 
 import (
 	"github.com/persistenceOne/persistenceSDK/schema/types"
@@ -8,25 +8,25 @@ import (
 
 func TestDuplicate(t *testing.T) {
 	type args struct {
-		properties types.Properties
+		propertyList []types.Property
 	}
 	tests := []struct {
 		name string
 		args args
 		want bool
 	}{
-		{"Positive Case, Unique Properties", args{properties: base.NewProperties(base.NewProperty(base.NewID("a"), base.NewFact(base.NewStringData("factA"))),
+		{"Positive Case, Unique Properties", args{propertyList: []types.Property{base.NewProperty(base.NewID("a"), base.NewFact(base.NewStringData("factA"))),
 			base.NewProperty(base.NewID("b"), base.NewFact(base.NewStringData("factB"))),
 			base.NewProperty(base.NewID("c"), base.NewFact(base.NewStringData("factC"))),
-			base.NewProperty(base.NewID("d"), base.NewFact(base.NewStringData("factD"))))}, false},
-		{"Negative Case, DuplicateExists", args{properties: base.NewProperties(base.NewProperty(base.NewID("a"), base.NewFact(base.NewStringData("factA"))),
+			base.NewProperty(base.NewID("d"), base.NewFact(base.NewStringData("factD")))}}, false},
+		{"Negative Case, DuplicateExists", args{propertyList: []types.Property{base.NewProperty(base.NewID("a"), base.NewFact(base.NewStringData("factA"))),
 			base.NewProperty(base.NewID("b"), base.NewFact(base.NewStringData("factB"))),
 			base.NewProperty(base.NewID("c"), base.NewFact(base.NewStringData("factC"))),
-			base.NewProperty(base.NewID("a"), base.NewFact(base.NewStringData("factD"))))}, true},
+			base.NewProperty(base.NewID("a"), base.NewFact(base.NewStringData("factD")))}}, true},
 	}
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			if got := Duplicate(tt.args.properties); got != tt.want {
+			if got := Duplicate(tt.args.propertyList); got != tt.want {
 				t.Errorf("Duplicate() = %v, want %v", got, tt.want)
 			}
 		})
